microservices/nlp-classifier/requirements.txt
joblib
numpy
pandas
transformers
scikit-learn
google-cloud-storage
torch